+++ Restobar/src/restobar/Models/Waiter.java
@@ -2,7 +2,7 @@ package restobar.Models;
 
 public class Waiter
 {
-    private long id;
+    private int id;
     private String name;
     private String lastName;
     //Constructors
@@ -12,17 +12,17 @@ public class Waiter
         this.name="";
         this.lastName="";
     }
-    public Waiter(long idWaiter,String name,String lastName)
+    public Waiter(int idWaiter,String name,String lastName)
     {
         this.id=id;
         this.name=name;
         this.lastName=lastName;
     }
     //Getters and setters
-    public long getId(){return this.id;}
+    public int getId(){return this.id;}
     public String getName(){return this.name;}
     public String getLastName(){return this.lastName;}
-    public void setId(long idWaiter){this.id=idWaiter;}
+    public void setId(int idWaiter){this.id=idWaiter;}
     public void setName(String name){this.name=name;}
     public void setLastName(String lastName){this.lastName=lastName;}
 }
